refactor: drop legacy booking API and update client
