+++ backend/app/api/v1/bookings.py
@@ -8,7 +8,7 @@ from app.schemas.api_booking import (
     BookingPublic,
     StripeSetupIntent,
 )
-from app.services import booking_service_v2
+from app.services import booking_service
 from app.dependencies import get_db
 
 router = APIRouter(prefix="/api/v1/bookings", tags=["bookings"])
@@ -18,7 +18,7 @@ async def create_booking_endpoint(
     payload: BookingCreateRequest, db: AsyncSession = Depends(get_db)
 ) -> BookingCreateResponse:
     try:
-        booking, client_secret = await booking_service_v2.create_booking(db, payload)
+        booking, client_secret = await booking_service.create_booking(db, payload)
     except ValueError as exc:
         raise HTTPException(status_code=400, detail=str(exc)) from exc
     booking_public = BookingPublic.model_validate(booking)
