+++ backend/tests/integration/test_driver_booking_actions_api.py
@@ -5,7 +5,7 @@ from datetime import datetime, timedelta, timezone
 import uuid
 
 from app.models.user_v2 import User, UserRole
-from app.models.booking_v2 import Booking, BookingStatus
+from app.models.booking import Booking, BookingStatus
 
 pytestmark = pytest.mark.asyncio
