+++ frontend/src/api-client/.openapi-generator/FILES
@@ -5,23 +5,19 @@ base.ts
 common.ts
 configuration.ts
 docs/AddressComponents.md
-docs/AppSchemasBookingBookingRead.md
-docs/AppSchemasBookingV2BookingRead.md
 docs/AuthApi.md
 docs/AvailabilityApi.md
 docs/AvailabilityResponse.md
 docs/AvailabilitySlotCreate.md
 docs/AvailabilitySlotRead.md
-docs/BookingCreate.md
 docs/BookingCreateRequest.md
 docs/BookingCreateResponse.md
 docs/BookingPublic.md
+docs/BookingRead.md
 docs/BookingSlot.md
 docs/BookingStatus.md
 docs/BookingStatusResponse.md
-docs/BookingUpdate.md
 docs/BookingsApi.md
-docs/CustomerBookingsApi.md
 docs/CustomerInfo.md
 docs/DriverBookingsApi.md
 docs/GeocodeApi.md
@@ -32,7 +28,6 @@ docs/HTTPValidationError.md
 docs/Location.md
 docs/LoginRequest.md
 docs/OAuth2Token.md
-docs/Price.md
 docs/RegisterRequest.md
 docs/RouteMetricsApi.md
 docs/SettingsApi.md
