--- frontend/src/api-client/docs/AppSchemasBookingBookingRead.md
@@ -1,35 +1,0 @@
-# AppSchemasBookingBookingRead
-
-Representation of a booking returned from API.
-
-## Properties
-
-Name | Type | Description | Notes
------------- | ------------- | ------------- | -------------
-**id** | **number** |  | [default to undefined]
-**user_id** | **number** |  | [default to undefined]
-**pickup_location** | **string** |  | [default to undefined]
-**dropoff_location** | **string** |  | [default to undefined]
-**time** | **string** |  | [default to undefined]
-**price** | **string** |  | [default to undefined]
-**status** | **string** |  | [default to undefined]
-**created_at** | **string** |  | [optional] [default to undefined]
-
-## Example
-
-```typescript
-import { AppSchemasBookingBookingRead } from './api';
-
-const instance: AppSchemasBookingBookingRead = {
-    id,
-    user_id,
-    pickup_location,
-    dropoff_location,
-    time,
-    price,
-    status,
-    created_at,
-};
-```
-
-[[Back to Model list]](../README.md#documentation-for-models) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to README]](../README.md)
