+++ frontend/src/components/ApiConfig.ts
@@ -1,5 +1,5 @@
 // Shared configuration and API client singletons.
-import { Configuration, AuthApi, BookingsApi, UsersApi, SetupApi, SettingsApi } from "@/api-client";
+import { Configuration, AuthApi, BookingsApi, DriverBookingsApi, UsersApi, SetupApi, SettingsApi } from "@/api-client";
 import { CONFIG } from "@/config";
 import { getAccessToken } from "@/services/tokenStore";
 
@@ -10,14 +10,15 @@ const configuration = new Configuration({
 });
 
 // Export **instances** (singletons)
-export const authApi     = new AuthApi(configuration);
-export const bookingsApi = new BookingsApi(configuration);
-export const usersApi    = new UsersApi(configuration);
-export const setupApi    = new SetupApi(configuration);
-export const settingsApi = new SettingsApi(configuration);
+export const authApi          = new AuthApi(configuration);
+export const bookingsApi      = new BookingsApi(configuration);
+export const driverBookingsApi = new DriverBookingsApi(configuration);
+export const usersApi         = new UsersApi(configuration);
+export const setupApi         = new SetupApi(configuration);
+export const settingsApi      = new SettingsApi(configuration);
 
 // (Keep if you still want the classes too)
-export { AuthApi, BookingsApi, UsersApi, SetupApi, SettingsApi } from "@/api-client";
+export { AuthApi, BookingsApi, DriverBookingsApi, UsersApi, SetupApi, SettingsApi } from "@/api-client";
 
 // Optional: token change hook
 // onTokenChange(() => { /* e.g. invalidate caches if needed */ });
