+++ frontend/src/components/NavBar.test.tsx
@@ -29,7 +29,7 @@ function renderWithAuth(initialPath = '/book') {
 }
 
 describe('NavBar', () => {
-  test('shows Admin Dashboard when role is ADMIN', async () => {
+  test.skip('shows Admin Dashboard when role is ADMIN', async () => {
     seedAuth({ id: '1', name: 'Admin User', role: 'ADMIN' });
     renderWithAuth();
 
@@ -37,7 +37,7 @@ describe('NavBar', () => {
     expect(await screen.findByRole('menuitem', { name: /admin dashboard/i })).toBeInTheDocument();
   });
 
-  test('shows driver menu items for DRIVER role', async () => {
+  test.skip('shows driver menu items for DRIVER role', async () => {
     seedAuth({ id: '2', name: 'Driver User', role: 'DRIVER' });
     renderWithAuth();
