+++ frontend/src/pages/Booking/RideDetailsPage.test.tsx
@@ -7,19 +7,24 @@ import { http, HttpResponse } from 'msw';
 import { server } from '@/__tests__/setup/msw.server';
 import { apiUrl } from '@/__tests__/setup/msw.handlers';
 import { AuthProvider } from '@/contexts/AuthContext';
+import { vi } from 'vitest';
+
+vi.mock('@/components/MapRoute', () => ({ MapRoute: () => <div /> }));
 
 test('Back button navigates to /history', async () => {
   const booking = {
-    id: 1,
-    pickup_location: 'A',
-    dropoff_location: 'B',
-    time: new Date().toISOString(),
-    price: 10,
-    status: 'completed',
+    id: '1',
+    pickup_address: 'A',
+    dropoff_address: 'B',
+    pickup_when: new Date().toISOString(),
+    estimated_price_cents: 1000,
+    deposit_required_cents: 500,
+    status: 'COMPLETED',
+    public_code: 'abc123',
   };
 
   server.use(
-    http.get(apiUrl('/bookings'), () => HttpResponse.json([booking]))
+    http.get(apiUrl('/api/v1/driver/bookings'), () => HttpResponse.json([booking]))
   );
 
   render(
@@ -44,17 +49,18 @@ test('Back button navigates to /history', async () => {
 
 test('shows tracking link when booking trackable', async () => {
   const booking = {
-    id: 1,
-    pickup_location: 'A',
-    dropoff_location: 'B',
-    time: new Date().toISOString(),
-    price: 10,
-    status: 'on_the_way',
+    id: '1',
+    pickup_address: 'A',
+    dropoff_address: 'B',
+    pickup_when: new Date().toISOString(),
+    estimated_price_cents: 1000,
+    deposit_required_cents: 500,
+    status: 'ON_THE_WAY',
     public_code: 'abc123',
   };
 
   server.use(
-    http.get(apiUrl('/bookings'), () => HttpResponse.json([booking]))
+    http.get(apiUrl('/api/v1/driver/bookings'), () => HttpResponse.json([booking]))
   );
 
   render(
